requirements.txt
transformers
torch
PyPDF2
pdfminer.six
beautifulsoup4
streamlit
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
python-dotenv
ipykernel
beautifulsoup4
faiss-cpu
requests


